requirements.txt
flask
flask-sqlalchemy
flask-security
flask-admin
Flask-Script
flask-wtf
Flask-Assets
WTForms-Alchemy
jsmin
uuid
sockjs-tornado
bcrypt
sockjsroom
tornado-redis
redis
flask-redis
guppy